pyproject.toml
[project]
name = "medchatguard"
version = "0.1.0"
description = ""
authors = [
    {name = "danieljames96",email = "danieljames96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "mlflow (>=2.21.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "huggingface-hub (>=0.30.2,<0.31.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "datasets (>=3.5.0,<4.0.0)",
    "transformers (>=4.51.1,<5.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "groq (>=0.24.0,<0.25.0)",
    "pyautogen (>=0.9,<0.10)",
    "chromadb (>=1.0.9,<2.0.0)",
    "langgraph (>=0.4.3,<0.5.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
